api_service/requirements.txt
django>=4.2,<5.0
djangorestframework
pandas
psycopg2-binary
django-filter